fix: remove prefetch logic + add nin
